[DIPU]support using torch profiler on nv (#559)
* support using torch profiler on nv

* modify tests according to comments

* fix test
